+++ src/main/kotlin/stickyWallet/commands/PayCommand.kt
@@ -164,7 +164,7 @@ class PayCommand : TabExecutor, UsePlugin {
         val temp = if (currency.decimalSupported) {
             amount.toBigDecimal()
         } else {
-            amount.toBigInteger().toBigDecimal()
+            amount.toBigDecimal().toBigInteger().toBigDecimal()
         }
         if (temp < BigDecimal.ZERO) throw NumberFormatException()
         temp
